Migrate TypeWriting component into layout file
