--- components/HomePage/TypeWrite.tsx
@@ -1,24 +1,0 @@
-/** @format */
-
-import React from "react";
-
-import Typewriter from "typewriter-effect";
-
-interface typeProps {
-  typePrompts: string[];
-}
-
-const TypeWrite = (props: typeProps) => {
-  return (
-    <Typewriter
-      options={{
-        strings: props.typePrompts,
-        autoStart: true,
-        loop: true,
-        deleteSpeed: 50,
-      }}
-    />
-  );
-};
-
-export default TypeWrite;
